+++ ai-services/groq.js
@@ -6,7 +6,7 @@ const configuration = c('groq', {
 
 const groq = new Groq(configuration);
 
-import AIService from '../ai-service.js';
+import AIService from '../tools/ai-service.js';
 
 import { replaceContent } from '../tools/censorship.js';
 import { mapAsyncIterator } from '../tools/map-async-iterator.js';
